⚙ Move `core/blockly` to `core/workspaces/blockly`
