+++ core/CHANGELOG.md
@@ -2,20 +2,20 @@
 
 ### 🀄 Blockly Workspace
 
-- 🆕 [v1.0.0] `blockly/parseInstructions`.
-- 🆕 [v1.0.0] `blockly/instanciateEveryBlock`.
-- 🆕 [v1.0.0] `blockly/blocks/defaultBlockParser`.
-- 🆕 [v1.0.0] `blockly/blocks/moveBuilder`.
+- 🆕 [v1.0.0] `parseInstructions`.
+- 🆕 [v1.0.0] `instanciateEveryBlock`.
+- 🆕 [v1.0.0] `blocks/defaultBlockParser`.
+- 🆕 [v1.0.0] `blocks/moveBuilder`.
 
 #### 🀄 Códimo Blocks
 
-- 🆕 [v1.0.0] `blockly/blocks/action_container`.
-- 🆕 [v1.0.0] `blockly/blocks/move_backward`.
-- 🆕 [v1.0.0] `blockly/blocks/move_forward`.
-- 🆕 [v1.0.0] `blockly/blocks/move_left`.
-- 🆕 [v1.0.0] `blockly/blocks/move_right`.
-- 🆕 [v1.0.0] `blockly/blocks/repeat_x`.
-- 🆕 [v1.0.0] `blockly/blocks/simple_loop`.
+- 🆕 [v1.0.0] `blocks/action_container`.
+- 🆕 [v1.0.0] `blocks/move_backward`.
+- 🆕 [v1.0.0] `blocks/move_forward`.
+- 🆕 [v1.0.0] `blocks/move_left`.
+- 🆕 [v1.0.0] `blocks/move_right`.
+- 🆕 [v1.0.0] `blocks/repeat_x`.
+- 🆕 [v1.0.0] `blocks/simple_loop`.
 
 ### 🤖 PixiJS Engine
