+++ core/workspaces/blockly/blocks/defaultBlockParser.js
