+++ core/workspaces/blockly/instanciateEveryBlock.js
